requirements/base.txt
coinbase==2.1.0
coinbase-commerce==1.0.1
django==3.2.4
django-ckeditor==6.5.1
django-coinbase==0.2
django-crispy-forms==1.14.0
django-currencies==0.10.1
django-environ==0.10.0
django-extensions==3.2.1
django-filter==22.1
django-fullurl==1.3
django-htmx==1.13.0
django-recaptcha==3.0.0
django-widget-tweaks==1.4.8
stripe==4.2.0
Pillow==9.5.0
whitenoise==6.4.0
django-redis==5.2.0
easy-thumbnails==2.8.5
django-allauth==0.54.0
crispy-bootstrap5==0.7